requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
segmentation-models-pytorch>=0.3.3

# Image Processing
Pillow>=9.0.0
numpy>=1.21.0

# Progress Bar
tqdm>=4.64.0

# Optional: For GeoTIFF support
# rasterio>=1.3.0

# Optional: For visualization
# matplotlib>=3.5.0

# Optional: For advanced augmentations
# albumentations>=1.3.0
